pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redclaw"
version = "2.0.0"
description = "RedClaw v2.0 — Enterprise Autonomous Penetration Testing Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "SparkStack Systems"}
]
keywords = ["pentesting", "security", "AI", "autonomous", "red-team"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Environment :: Console",
]

dependencies = [
    "paramiko>=3.4.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.43",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "requests>=2.31.0",
    "msgpack>=1.0.7",
    "psutil>=5.9.8",
    "aiohttp>=3.9.3",
    "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

# ── CLI Entry Points ──────────────────────────────────────────────────────────
# After `pip install redclaw`, these commands are globally available:
#   redclaw          →  main CLI (interactive or subcommands)
#   redclaw-doctor   →  quick health check shortcut
#   redclaw-skin     →  launch Claude Code skin mode
[project.scripts]
redclaw = "redclaw.cli.app:main"
redclaw-doctor = "redclaw.cli.shortcuts:doctor"
redclaw-skin = "redclaw.cli.shortcuts:skin"

[project.urls]
Homepage = "https://github.com/sparkstack/redclaw"
Documentation = "https://github.com/sparkstack/redclaw/docs"
Repository = "https://github.com/sparkstack/redclaw"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
